--- LibDS/examples/SimpleDS/build.bat
@@ -1,3 +1,0 @@
-nmake -clean
-qmake -config release
-nmake
