+++ LightningDS/build.bat
@@ -1,0 +1,3 @@
+nmake -clean
+qmake -config release
+nmake
