+++ LightningDS/resources/resources.qrc
@@ -1,0 +1,7 @@
+<RCC>
+    <qresource prefix="/pixmaps">
+	<file>lightning.png</file>
+        <file>frc-logo.png</file>
+        <file>joystick.png</file>
+    </qresource>
+</RCC>
